fixes plugin list with version-constrained plugins (#3)
if a plugin is installed with a specific version, then that version
constraint will pop up in a "ui" element. this change guards against
trying to extract the matches when the regex fails and returns nil.

a better implementation would probably NOT use the "ui" type and
perform a more involved parsing of the individual name and version
fields.